+++ hahomematic/platforms/device.py
@@ -14,6 +14,7 @@ from typing import Any, Final
 import orjson
 
 from hahomematic import central as hmcu
+from hahomematic.async_support import loop_check
 from hahomematic.const import (
     DEFAULT_DEVICE_DESCRIPTIONS_DIR,
     DEFAULT_PARAMSET_DESCRIPTIONS_DIR,
@@ -45,13 +46,7 @@ from hahomematic.platforms.event import GenericEvent
 from hahomematic.platforms.generic.entity import GenericEntity
 from hahomematic.platforms.support import PayloadMixin, get_device_name
 from hahomematic.platforms.update import HmUpdate
-from hahomematic.support import (
-    CacheEntry,
-    Channel,
-    check_or_create_directory,
-    loop_check,
-    reduce_args,
-)
+from hahomematic.support import CacheEntry, Channel, check_or_create_directory, reduce_args
 
 _LOGGER: Final = logging.getLogger(__name__)
 
@@ -546,7 +541,7 @@ class HmDevice(PayloadMixin):
                 await self.central.refresh_firmware_data(device_address=self._device_address)
 
         if refresh_after_update_intervals:
-            self.central.create_task(target=refresh_data(), name="refresh_firmware_data")
+            self.central.looper.create_task(target=refresh_data(), name="refresh_firmware_data")
 
         return update_result
 
@@ -854,4 +849,6 @@ class _DefinitionExporter:
                 )
             return DataOperationResult.SAVE_SUCCESS
 
-        return await self._central.async_add_executor_job(_save)
+        return await self._central.looper.async_add_executor_job(
+            _save, name="save-device-description"
+        )