+++ hahomematic/platforms/event.py
@@ -7,6 +7,7 @@ import logging
 from typing import Any, Final
 
 from hahomematic import support as hms
+from hahomematic.async_support import loop_check
 from hahomematic.const import (
     CLICK_EVENTS,
     DEVICE_ERROR_EVENTS,
@@ -23,7 +24,6 @@ from hahomematic.platforms import device as hmd
 from hahomematic.platforms.decorators import config_property
 from hahomematic.platforms.entity import BaseParameterEntity
 from hahomematic.platforms.support import EntityNameData, generate_unique_id, get_event_name
-from hahomematic.support import loop_check
 
 _LOGGER: Final = logging.getLogger(__name__)