+++ changelog.md
@@ -1,6 +1,8 @@
 # Version 2024.4.8 (2024-04-13)
 
 - Make entity event async
+- Extract looper from central and reuse for json/xml_rpc
+- Move loop_check to async_support
 
 # Version 2024.4.7 (2024-04-13)